[Bronze III] Title: 수학은 체육과목 입니다, Time: 128 ms, Memory: 14200 KB -BaekjoonHub

+++ 백준/Bronze/15894. 수학은 체육과목 입니다/수학은 체육과목 입니다.java
@@ -4,9 +4,8 @@ public class Main {
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
-		// N의 범위가 int의 범위를 초과하므로 long형을 써서 N을 입력 받음
-		long n = Integer.parseInt(br.readLine());
-		
+		// N의 범위가 int의 범위를 초과하므로 long형을 써서 N을 입력받음
+		long n = Long.parseLong(br.readLine());
 		System.out.println(n * 4);
 	}
 }
